pyproject.toml
[tool.poetry]
name = "crewai-snippet"
version = "0.1.0"
description = ""
authors = ["csycp frontend@acalab.co.kr"]
readme = "README.md"
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.10, <=3.13"
crewai = {extras = ["tools"], version = "^0.30.11"}


python-dotenv = "^1.0.1"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
start = 'src.main:main'
